requirements.txt
kaggle==1.6.17
matplotlib==3.9.2
mlflow==2.17.0
mlflow_skinny==2.17.0
numpy==1.23.5
opencv_python==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.5.2
scipy==1.14.1
setuptools==44.0.0
skimage==0.0
streamlit==1.39.0
tensorflow==2.17.0
azureml-core
azureml-mlflow